Codigo, MNIST + corrección errores
Se añade la base de datos MNIST y se arreglan algunos errores. A partir de aquí el comando "make" puede que no funcione, pues se comenzará la verificación y posibles cambios en la capa "fully_connected". En su lugar, usar el comando en el fichero TFG/Codigo/CNN/capas,/fully_connected/compilar

+++ Codigo/CNN/capas/flatten/flatten.h
@@ -10,7 +10,7 @@ class Flatten
         int canales;
         int filas;
         int cols;
-        int max;
+        float max;
 
     public:
         Flatten(const vector<vector<vector<float>>> &input);
